test on use border in Analysis
