+++ test_autolens/pipeline/phase/dataset/test_phase_dataset.py
@@ -17,124 +17,6 @@ pytestmark = pytest.mark.filterwarnings(
 directory = path.dirname(path.realpath(__file__))
 
 
-class TestMakeAnalysis:
-    def test__auto_einstein_radius_is_used__einstein_radius_used_in_analysis(
-        self, imaging_7x7, mask_7x7
-    ):
-        # If position threshold is input (not None) and positions are input, make the positions part of the lens dataset.
-
-        phase_imaging_7x7 = al.PhaseImaging(
-            search=mock.MockSearch("test_phase"),
-            settings=al.SettingsPhaseImaging(
-                settings_lens=al.SettingsLens(auto_einstein_radius_factor=None)
-            ),
-        )
-        tracer = mock.MockTracer(einstein_radius=2.0)
-
-        phase_imaging_7x7.modify_settings(
-            dataset=imaging_7x7,
-            results=mock.MockResults(max_log_likelihood_tracer=tracer),
-        )
-
-        analysis = phase_imaging_7x7.make_analysis(
-            dataset=imaging_7x7, mask=mask_7x7, results=mock.MockResults()
-        )
-
-        assert analysis.settings.settings_lens.einstein_radius_estimate == None
-
-        phase_imaging_7x7 = al.PhaseImaging(
-            search=mock.MockSearch("test_phase"),
-            settings=al.SettingsPhaseImaging(
-                settings_lens=al.SettingsLens(auto_einstein_radius_factor=1.0)
-            ),
-        )
-
-        tracer = mock.MockTracer(einstein_radius=2.0)
-
-        phase_imaging_7x7.modify_settings(
-            dataset=imaging_7x7,
-            results=mock.MockResults(max_log_likelihood_tracer=tracer),
-        )
-
-        analysis = phase_imaging_7x7.make_analysis(
-            dataset=imaging_7x7,
-            mask=mask_7x7,
-            results=mock.MockResults(max_log_likelihood_tracer=tracer),
-        )
-
-        assert analysis.settings.settings_lens.einstein_radius_estimate == 2.0
-
-    def test__use_border__determines_if_border_pixel_relocation_is_used(
-        self, imaging_7x7, mask_7x7
-    ):
-        # noinspection PyTypeChecker
-
-        lens_galaxy = al.Galaxy(
-            redshift=0.5, mass=al.mp.SphericalIsothermal(einstein_radius=100.0)
-        )
-        source_galaxy = al.Galaxy(
-            redshift=1.0,
-            pixelization=al.pix.Rectangular(shape=(3, 3)),
-            regularization=al.reg.Constant(coefficient=1.0),
-        )
-
-        phase_imaging_7x7 = al.PhaseImaging(
-            galaxies=[lens_galaxy, source_galaxy],
-            settings=al.SettingsPhaseImaging(
-                settings_masked_imaging=al.SettingsMaskedImaging(
-                    grid_inversion_class=al.Grid2D
-                ),
-                settings_pixelization=al.SettingsPixelization(use_border=True),
-            ),
-            search=mock.MockSearch("test_phase"),
-        )
-
-        analysis = phase_imaging_7x7.make_analysis(
-            dataset=imaging_7x7, mask=mask_7x7, results=mock.MockResults()
-        )
-        analysis.masked_dataset.grid_inversion[4] = np.array([[500.0, 0.0]])
-
-        instance = phase_imaging_7x7.model.instance_from_unit_vector([])
-        tracer = analysis.tracer_for_instance(instance=instance)
-        fit = analysis.masked_imaging_fit_for_tracer(
-            tracer=tracer, hyper_image_sky=None, hyper_background_noise=None
-        )
-
-        assert fit.inversion.mapper.source_grid_slim[4][0] == pytest.approx(
-            97.19584, 1.0e-2
-        )
-        assert fit.inversion.mapper.source_grid_slim[4][1] == pytest.approx(
-            -3.699999, 1.0e-2
-        )
-
-        phase_imaging_7x7 = al.PhaseImaging(
-            galaxies=[lens_galaxy, source_galaxy],
-            settings=al.SettingsPhaseImaging(
-                settings_masked_imaging=al.SettingsMaskedImaging(
-                    grid_inversion_class=al.Grid2D
-                ),
-                settings_pixelization=al.SettingsPixelization(use_border=False),
-            ),
-            search=mock.MockSearch("test_phase"),
-        )
-
-        analysis = phase_imaging_7x7.make_analysis(
-            dataset=imaging_7x7, mask=mask_7x7, results=mock.MockResults()
-        )
-
-        analysis.masked_dataset.grid_inversion[4] = np.array([300.0, 0.0])
-
-        instance = phase_imaging_7x7.model.instance_from_unit_vector([])
-        tracer = analysis.tracer_for_instance(instance=instance)
-        fit = analysis.masked_imaging_fit_for_tracer(
-            tracer=tracer, hyper_image_sky=None, hyper_background_noise=None
-        )
-
-        assert fit.inversion.mapper.source_grid_slim[4][0] == pytest.approx(
-            200.0, 1.0e-4
-        )
-
-
 class TestAutoPositions:
     def test__updates_correct_using_factor(
         self, imaging_7x7, image_7x7, noise_map_7x7, mask_7x7
